requirements-gpu.txt
tensorflow-gpu==2.0.0-beta1
pandas
gradient-statsd
scikit-learn
jupyterlab
matplotlib
gradient
gsutil
click
rope